+++ tests/core/test_initialization.py
@@ -180,22 +180,6 @@ class ProjectSettingsTest(unittest.TestCase):
         os.environ.pop(envkey)
         self.assertEqual(value, 'kinto_redis.storage')
 
-    def test_can_continue_to_use_cliquet_names(self):
-        settings = {
-            'kinto.permission_backend': 'cliquet.permission.memory'
-        }
-        with mock.patch('kinto.core.initialization.logger.warn') as mocked:
-            new_settings = self.settings(settings)
-            warning_message = ''.join([
-                "Backend settings referring to cliquet are DEPRECATED. ",
-                "Please update your kinto.permission_backend setting to ",
-                "kinto.core.permission.memory ",
-                "(was: cliquet.permission.memory).",
-                ])
-            mocked.assert_called_once_with(warning_message)
-            self.assertEqual(new_settings['permission_backend'],
-                             'kinto.core.permission.memory')
-
 
 class ApplicationWrapperTest(unittest.TestCase):